Complete merge of refactor-arq7 branch
- Refactored large arq7.rs file into smaller, focused modules
- Created new modules: backup_config, backup_folder, backup_folders, backup_plan, backup_record, backup_set, blob_loc, encrypted_keyset, node, virtual_fs, utils
- Moved common utility functions (encryption helpers) to utils module
- Fixed compilation errors and import issues
- All tests passing
