+++ arq/src/arq7/backup_plan.rs
@@ -1,8 +1,9 @@
-use super::backup_folders::load_json_with_encryption;
 use super::encrypted_keyset::EncryptedKeySet;
+use super::utils::load_json_with_encryption;
 use crate::error::Result;
-use serde::{de::Deserializer, Deserialize, Serialize};
+use serde::{de::Deserializer, Deserialize};
 use std::collections::HashMap;
+
 use std::path::Path;
 
 /// TransferRate configuration for backup plans
@@ -345,27 +346,4 @@ impl BackupPlan {
     ) -> Result<BackupPlan> {
         load_json_with_encryption(path, keyset)
     }
-}
-
-/// Helper function to load JSON from either encrypted or unencrypted file
-fn load_json_with_encryption<T, P>(path: P, keyset: Option<&EncryptedKeySet>) -> Result<T>
-where
-    T: for<'de> serde::Deserialize<'de>,
-    P: AsRef<Path>,
-{
-    let path_ref = path.as_ref();
-
-    if let Some(keyset) = keyset {
-        // Check if file is encrypted
-        if is_file_encrypted(path_ref)? {
-            // Decrypt and parse
-            let json_content = decrypt_json_file(path_ref, keyset)?;
-            return Ok(serde_json::from_str(&json_content)?);
-        }
-    }
-
-    // Load as regular unencrypted file
-    let file = File::open(path_ref)?;
-    let reader = BufReader::new(file);
-    return Ok(serde_json::from_reader(reader)?);
 }
