+++ arq/src/arq7/encrypted_keyset.rs
@@ -3,7 +3,7 @@ use crate::object_encryption::calculate_hmacsha256;
 use crate::type_utils::ArqRead;
 use byteorder::ReadBytesExt;
 use std::fs::File;
-use std::io::{BufRead, Read, Seek};
+use std::io::{BufRead, Seek};
 use std::path::Path;
 
 /// EncryptedKeySet represents the encryptedkeyset.dat file
@@ -51,9 +51,7 @@ const ENCRYPTED_KEYSET_HEADER: [u8; 25] = [
 impl EncryptedKeySet {
     pub fn from_master_keys(master_keys: Vec<Vec<u8>>) -> Result<Self> {
         if master_keys.len() != 3 {
-            return Err(Error::InvalidFormat(
-                "Expected 3 master keys".to_string(),
-            ));
+            return Err(Error::InvalidFormat("Expected 3 master keys".to_string()));
         }
         Ok(EncryptedKeySet {
             encryption_key: master_keys[0].clone(),
